stripping out all non-core code
