--- demba/extract_features_v3.py
@@ -1,73 +1,0 @@
-import numpy as np
-import pandas as pd
-from pathlib import Path
-import cv2
-from itertools import combinations, permutations
-from demba.utils.roi_utils import estimate_roi
-idx = pd.IndexSlice
-
-class FeatureExtractor:
-
-    def __init__(self, config_path, video_path):
-        self.config_path, self.video_path = config_path, video_path
-        self.h5_path = str(next(Path(self.video_path).parent.glob(Path(self.video_path).stem + '*_filtered.h5')))
-        self._load_pose_data()
-        self._estimate_roi()
-        self._calc_all_features()
-
-    def _load_pose_data(self):
-        self.pose_df = pd.read_hdf(self.h5_path)
-        scorer = self.pose_df.columns.get_level_values(0)[0]
-        self.pose_df = self.pose_df.loc[:, scorer]
-        self.pose_df = self.pose_df.loc[:, idx[:, :, ('x', 'y')]]
-        self.pose_df.columns = self.pose_df.columns.remove_unused_levels()
-        self.pose_df = self.pose_df.sort_index(axis=1)
-        self.individuals, self.bodyparts = [list(self.pose_df.columns.levels[i]) for i in [0, 1]]
-
-    def _estimate_roi(self):
-        cap = cv2.VideoCapture(self.video_path)
-        ret, frame = cap.read()
-        if not ret:
-            print(f'could not extract a reference frame from {Path(self.video_path).name}')
-            return
-        frame = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
-        roi_vis_path = str(Path(self.video_path).parent / (Path(self.video_path).stem + '_roi.png'))
-        self.roi_x, self.roi_y, self.roi_r = estimate_roi(frame, output_path=roi_vis_path)
-        self.frame_height, self.frame_width = frame.shape[:-1]
-        cap.release()
-
-    def _calc_all_features(self):
-        self._calc_intermdiates()
-        self.feature_matrix = []
-        self.feature_matrix.append(self._calc_nfish_frame())
-        self.feature_matrix.append(self._calc_nfish_pipe())
-        self.feature_matrix = pd.concat(self.feature_matrix, axis=1)
-
-    def _calc_nfish_frame(self):
-        nfish_frame = self.pose_df.groupby('individuals', axis=1).any().sum(axis=1)
-        nfish_frame.name = 'nfish_frame'
-        return nfish_frame
-
-    def _calc_nfish_pipe(self):
-        tmp_df = self.pose_df.loc[:, idx[:, 'stripe1', :]]
-        tmp_df.loc[:, idx[:, :, 'x']] -= self.roi_x
-        tmp_df.loc[:, idx[:, :, 'y']] -= self.roi_y
-        tmp_df = (tmp_df ** 2).groupby('individuals', axis=1).sum(min_count=2) ** 0.5
-        nfish_pipe = (tmp_df <= self.roi_r).sum(axis=1)
-        nfish_pipe.name = 'nfish_pipe'
-        return nfish_pipe
-
-    def _calc_min_dist_nose_to_stripe4(self):
-        candidate_dists = []
-        for id1, id2 in list(permutations(self.individuals, 2)):
-            dists = self.pose_df.loc[:, idx[id1, 'nose', :]].values - self.pose_df.loc[:, idx[id2, 'stripe4', :]].values
-            candidate_dists.append(np.hypot(dists[:, 0], dists[:, 1]))
-        return pd.Series(np.nanmin(np.vstack(candidate_dists), axis=0), name='min_dist_nose_to_stripe4')
-
-
-
-
-
-
-
-
