+++ main.py
@@ -1,13 +1,24 @@
-import os
-from demba.extract_features_v3 import FeatureExtractor
+from pathlib import Path
+from demba.extract_features import FeatureExtractor
+from demba.estimate_pose import estimate_pose
 
 
+def run_analysis(analysis_targets, dlc_config, quivering_annotations, run_pose_estimation=True, featurize=True):
+    for target in analysis_targets:
+        video_path = analysis_dir_path / 'Videos' / target / f'{target}.mp4'
+        if not video_path.exists():
+            print(f'File Not Found: {video_path}. Skipping')
+            continue
+        if run_pose_estimation:
+            print(f'running pose estimation on {video_path.name}')
+            estimate_pose(dlc_config, video_path, visualize=True)
+        if featurize:
+            fe = FeatureExtractor(video_path, quivering_annotations)
+            fe.extract_all_features()
 
-project = 'demasoni_singlenuc'
-prefix = 'C:\\Users\\tucke\\DLC_Projects'
-project_path = os.path.join(prefix, project)
-config = os.path.join(project_path, 'config.yaml')
-shuffle = 1
 
-video = os.path.join(project_path, 'testclip\\testclip.mp4')
-featurizer = FeatureExtractor(config, video)
+dlc_config_path = Path('/home/tlancaster/DLC/demasoni_singlenuc/config.yaml')
+analysis_dir_path = dlc_config_path.parent / 'Analysis'
+quivering_annotation_path = analysis_dir_path / 'Annotations' / 'Mbuna_behavior_annotations.xlsx'
+analysis_target_list = ['BHVE_group1_097200-098999']
+run_analysis(analysis_target_list, dlc_config_path, quivering_annotation_path, True, True)
